requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# Google AI integration
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Vector database
langchain-milvus>=0.1.0
pymilvus>=2.3.0

# ML/AI packages
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.65.0

# Image processing and downloading
requests>=2.31.0
Pillow>=9.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# For future API development
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Development and testing
pytest>=7.4.0
requests>=2.31.0
